app controller cleanup works now
When a space is deleted, delete associated resources (including
namespace) and wait for corresponding app controller to stop.

During the ensuing decco controller shutdown, call Stop() on the
remaining app controllers to shut them down without deleting
their backing resources, so that they can be restarted as children
of the future decco controller.