add explicit metrics containerPort 10254 to nginx-ingress app
This is needed for Prometheus scraping.
See Slack conversation for details:
https://platform9.slack.com/archives/C9QQ23ML5/p1566334371071300

I first attempted a more general solution to this: automatically
inserting containerPort entries into the deployment's pod template
from Endpoint entries defined in the decco App. This turned out
to be more complicated than thought, because an Endpoint doesn't
specify the container in the pod that exposes the endpoint, and
a pod can have multiple containers.

So for now, just add a hard-coded 10254 containerPort to the
nginx-ingress' pod template spec. We'll revisit this if a more
general solution is required later.