don't create default-http deployment and ingress by default to allow user-managed root path handler
Now that the Serenity app adds the 'ui-redirect' endpoint for the '/'
path, decco should not create the 'http-ingress' (misnamed, it should
have been called 'default-http') ingress anymore,
because it uses the same path and therefore conflicts.
The default-http deployment is also useless, and can be removed by default.
However, the 'default-http' service needs to stay, because the
nginx-ingress controller requires a default service specifier. It
can be any dummy service, even without backing pods.